Move implementation of Head/Main/NextScript
